requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
torch
ultralytics
opencv-python-headless
Pillow
requests
pytesseract
gunicorn
